test: split database, resource integration tests
